Add a Pause of 20 ms between every packet sent
This loss caused delay from the moment the user pressed the Stop button and the exact moment the mjUA stopped to reproduce the audio (RTP Packets) it received. Now with the same mjUA elaboration delay the audio stop almost instantly. The mjUA so have no more old packets to elaborate or reproduce and receive a right amount of RTP Packet to reproduce till the exact moment the User press the Stop button.