Add Listen4Close Runnable Class
This class is instantiated, inserted in a thread and started once the useragent receives the 200OK message and send the Ack Request, so the RTP flow begins. This thread simply listen for any new message from Bob on the port 5070, ready to close the call in case Bob send a BYE message.